refactor: CarList 클래스 추가
Game 클래스에서 바로 우승자를 구하는 등의 로직을 처리하지 않고 Car들을 모아서 리스트로 관리할 컬렉션 CarList를 만들어서 리팩토링. Car와 CarList 클래스를 domain 폴더로 이동
