+++ README.md
@@ -15,12 +15,15 @@
 ## 구현할 클래스 리스트
 - Game
   - 게임 진행을 담당하는 클래스
-  - 자동차를 생성하고 게임 실행 횟수만큼 게임 진행
+  - 이름과 게임 실행 횟수를 입력받고 게임 실행 횟수만큼 게임 진행
 - Car
   - 자동차 클래스
   - 이름, 위치를 필드로 가짐
   - 무작위 수를 생성하고 조건에 맞으면 전진
   - 게임의 각 회차마다 본인의 이름과 위치를 출력
+- CarList
+  - Car들을 모아놓은 컬렉션
+  - 자동차를 생성하고 우승자를 구하는 등 자동차들에 관련된 로직 담당
 - InputHandler
   - 입력을 담당할 클래스
 - Validator
