+++ src/test/java/racingcar/GameTest.java
@@ -11,6 +11,10 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import racingcar.Game;
+import racingcar.domain.Car;
+import racingcar.domain.CarList;
+
 class GameTest {
 
 	Game game;
@@ -28,6 +32,9 @@ class GameTest {
 		Field carListField = Game.class.getDeclaredField("carList");
 		carListField.setAccessible(true);
 
+		Field listField = CarList.class.getDeclaredField("list");
+		listField.setAccessible(true);
+
 		// given
 		String input = "test1, test2";
 		InputStream in = new ByteArrayInputStream(input.getBytes());
@@ -37,8 +44,9 @@ class GameTest {
 		inputName.invoke(game);
 
 		// then
-		List<Car> carList = (List<Car>)carListField.get(game);
-		assertThat(carList.size()).isEqualTo(2);
+		CarList carList = (CarList)carListField.get(game);
+		List<Car> list = (List<Car>)listField.get(carList);
+		assertThat(list.size()).isEqualTo(2);
 	}
 
 	@Test
