+++ src/test/java/racingcar/domain/CarTest.java
@@ -1,7 +1,6 @@
-package racingcar;
+package racingcar.domain;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
 
 import java.io.ByteArrayOutputStream;
 import java.io.OutputStream;
@@ -11,6 +10,8 @@ import java.util.Collections;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
+import racingcar.domain.Car;
 
 class CarTest {
 	Car car;
